+++ confignode/src/main/java/org/apache/iotdb/confignode/conf/ConfigNodeConstant.java
@@ -28,25 +28,19 @@ public class ConfigNodeConstant {
           : "UNKNOWN";
 
   public static final String GLOBAL_NAME = "IoTDB ConfigNode";
-
-  public static final String CONFIGNODE_PACKAGE = "org.apache.iotdb.confignode.service";
-  public static final String JMX_TYPE = "type";
-
   public static final String CONFIGNODE_CONF = "CONFIGNODE_CONF";
-
   public static final String CONFIGNODE_HOME = "CONFIGNODE_HOME";
 
   public static final String ENV_FILE_NAME = "confignode-env";
+  public static final String CONF_NAME = "iotdb-confignode.properties";
+  public static final String SPECIAL_CONF_NAME = "iotdb-confignode-special.properties";
 
+  public static final String CONFIGNODE_PACKAGE = "org.apache.iotdb.confignode.service";
+  public static final String JMX_TYPE = "type";
   public static final String CONFIGNODE_JMX_PORT = "confignode.jmx.port";
 
   public static final String DATA_DIR = "data";
-
   public static final String CONF_DIR = "conf";
-
-  public static final String CONF_NAME = "iotdb-confignode.properties";
-
-  public static final String SPECIAL_CONF_NAME = "iotdb-confignode-special.properties";
 
   public static final int MIN_SUPPORTED_JDK_VERSION = 8;
