+++ confignode/src/main/java/org/apache/iotdb/confignode/service/register/JMXService.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.confignode.service;
+package org.apache.iotdb.confignode.service.register;
 
 import org.apache.iotdb.confignode.conf.ConfigNodeConstant;
